modifipic_app/requirements.txt
asgiref==3.2.3
backcall==0.1.0
decorator==4.4.1
Django==3.0.3
djangorestframework==3.11.0
entrypoints==0.3
flake8==3.7.9
ipython==7.12.0
ipython-genutils==0.2.0
jedi==0.16.0
mccabe==0.6.1
numpy==1.18.1
opencv-python==4.2.0.32
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==3.0.3
psycopg2-binary==2.8.4
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.5.2
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
traitlets==4.3.3
wcwidth==0.1.8
